fix(detector): detector, tollg and loc csv, moved to pkg for cli tool detector reuse
