+++ pkg/detector/bbox.go
