+++ pkg/detector/bbox_test.go
