+++ pkg/detector/detector.go
