+++ pkg/detector/detector_test.go
