+++ pkg/detector/line_test.go
