+++ pkg/detector/storage.go
