+++ pkg/detector/storage_test.go
