+++ pkg/location/csv/testdata/coopdrive-gps-pings-2022.04.06.csv
